Implemented a projectile class
Actions have a projectile attached to them and will fire it when the action is performed The effects are done to the target when it hits.
Currently only made the effect for magic projectile.
Going to add a class alternative for melee actions. Could just use projectile with an option for no projectile but that doesnt really make sense.
Also got rid of old non-hex nodes.