Fixed a ton of status related bugs
Lots of exceptions relating to removing the status when dying, turns not ending when dying, removing and applying statuses should all be fixed now. :thumbsup:

Smoothed out unit death so that the unit graphics will be destroyed, any particle systems will fade out and then the unit will be destroyed.

Made status particle systems follow the unit and stack in intensity when you have more of the same effect.

Shooting a node now aims at the top rather than the center of the node.