Changes and fixes to statuses
Statuses are now checked at impact (initial effects only) as well as at the top of the units turn instead of the game turn. After the effect is applied the turn will continue. At the end of the turn the effects will be checked for 0 duration, if so they are removed.

Fixed bug when unit died from a status at the start of the turn

Added a new status. It has no original effects yet

Fixed bug when reading in more than one status

Small changes to card UI. Need to make fresh new cards at some point

Fixed bugs when enemies ran out of targets