Added ninja :skull: :star:
Ninja has 2 abilities. A copy of throw rock that does one more damage and has lower initiative (doesn't have its own effects yet) and a wind buff that gives himself 2 more movespeed and reduces incoming damage by 1 for 1 turn (has effect (is VERY cool))

Added a status undo-er for stat changes at the end of the units round so that when the effect is gone, the stat modifications are cleaned up as well.

Made the game wait for projectiles to finish before fully ending the units turn so that statuses will last right until the round is full over. This might cause bugs but it seems fine at the moment.

Imported some new particle effects to use later.

Fixed a bug with AI trying to determine empty status fitness. I don't know how this keeps happening but it does and it should be fixed now so :unamused: .