Overhauling AI 32.8% complete :bar_chart:
Actions should now be valued on fitness alone rather than number of valid targets.

Split possible actions into its own file because it was way too big.

The goal at the end is for enemies to value actions based on a better fitness rating, choose nodes with no units as the target in the case of AOEs if that is optimal and choose their nodes to move to more carefully (running away, staying further away from threats).