Enemies can run away :running:
Enemies will attempt to move to the maximum range of their action when using it, rather than rushing towards enemies with every move.

Enemies will slightly value moving towards the player when using 0 range moves. This should limit them from stalling or running away for no reason, not that that was a big problem yet.

Added functionality to test the fitness of targetting nodes that are empty, but in range for AOE and it worked correctly however it wasn't optimized and so it was very slow. I can have it remember what fitness score a node has when it's first checked which should speed it up enough to be usable. Will do that later.